Update insert1.php
This is the codes that records the QR code inputs into my database and shows them in the table in the front page of my webpage. When I try to add something to it, my system wont work. I also want to record 'TIME OUT' and 'STATUS' .

My Database name:                                                                   qr code database
The name of table that stores data in 'qr code database:          table_attendance